修正bug
寫入sql型別錯誤
點選變更成分時未更新正確group

+++ WpfApp1/DataClass/Entity/FabricProportion.cs
@@ -11,7 +11,7 @@ namespace WpfApp1.DataClass.Entity
         public int ProportionNo { get; set; }
         public int ColorNo { get; set; }
         public int YarnPriceNo { get; set; }
-        public float Proportion { get; set; }
+        public decimal Proportion { get; set; }
         public int Group { get; set; }
         public DateTime CreateDate { get; set; }
         public DateTime UpdateDate { get; set; }
