+++ WpfApp1/DataClass/Entity/ProcessSequence.cs
@@ -14,7 +14,7 @@ namespace WpfApp1.DataClass.Entity
         public int FabricID { get; set; }
         public string ColorNoString { get; set; }
         public ProcessItem ProcessItem { get; set; }
-        public float Loss { get; set; }
+        public decimal Loss { get; set; }
         public int WorkPay { get; set; }
         public int Order { get; set; }
         public int Group { get; set; }
